feat: native MCP integration via SSE
