--- mcp/.gitignore
@@ -1,4 +1,0 @@
-node_modules/
-build/
-*.log
-.DS_Store
