--- mcp/package.json
@@ -1,23 +1,0 @@
-{
-    "name": "nudlers-mcp",
-    "version": "1.0.0",
-    "description": "MCP server for Nudlers personal finance application",
-    "type": "module",
-    "main": "build/index.js",
-    "scripts": {
-        "build": "tsc",
-        "start": "node build/index.js",
-        "dev": "tsc --watch"
-    },
-    "dependencies": {
-        "@modelcontextprotocol/sdk": "^1.0.0",
-        "zod": "^3.22.4"
-    },
-    "devDependencies": {
-        "@types/node": "^20.10.0",
-        "typescript": "^5.3.0"
-    },
-    "engines": {
-        "node": ">=18.0.0"
-    }
-}
